outsourced interface
